Added dumb test app for triangle detection
