+++ VS/HaloAimBot/HaloAimBot.vcxproj
@@ -158,6 +158,7 @@
     <ClCompile Include="RedTriangleHunter.cpp" />
     <ClCompile Include="TargetTracker.cpp" />
     <ClCompile Include="TargetTrackingTest.cpp" />
+    <ClCompile Include="TriangleFinderTest.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="MotionModel.h" />
