+++ VS/HaloAimBot/HaloAimBot.vcxproj.filters
@@ -35,6 +35,9 @@
     <ClCompile Include="TargetTracker.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="TriangleFinderTest.cpp">
+      <Filter>Source Files\Tests</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Processing.h">
